5 задача
